requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.10.0
tensorflow>=2.8.0
matplotlib>=3.4.0
seaborn>=0.11.0
shap>=0.41.0
lime>=0.2.0
fairlearn>=0.7.0
aif360>=0.4.0
pyyaml>=5.4.0
pytest>=7.0.0
joblib>=1.1.0
scipy>=1.7.0
cryptography>=36.0.0

